refactor(atlassian): move pack tests out of the runtime payload
Same relocation as converters, applied to the atlassian pack's five skills.
`jira-team-status` was already in a `tests/` directory, but under `.apm/` —
the boundary is the runtime export path, not the directory name.

The SSO trios get an in-step dependency probe. Both `importorskip("credbroker")`
at module scope, so if a repoint ever detaches them from the
`pip install credbroker[crypto]` twenty-odd lines above, pytest exits 0 with the
whole file skipped and the cookie-confinement and fail-closed-config assertions
stop running with CI still green. A comment would not survive a reorder; the
probe does.

`_DUPLICATED` in `tools/test-lint-sso-config.py` splits rather than moves. It
pins five files byte-identical across jira and confluence-crawler, and that set
now spans the boundary: the two loaders stay under `scripts/`, the three tests
move. Both halves keep a parity check — a one-sided edit to a duplicated test is
how a one-sided edit to the control it guards goes unnoticed.

`tools/check-atlassian-phase3-readiness.py` invokes one of these suites by
absolute path and reports `fail` when it is missing, so it is repointed in the
same commit.

Before/after per suite: 45 / 9 / 48 / 9 / 51 passed, four exit-code harnesses at
9 checks each — identical on both sides.

Refs: docs/specs/pack-test-boundary-remaining-packs/spec.md (AC1, AC2, AC6, AC6a, AC6c)
