+++ packs/atlassian/tests/skills/confluence-crawler/test_auth_selector.py
