+++ packs/atlassian/tests/skills/confluence-crawler/test_exit_codes.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python3
-"""Smoke test for jira_align.py's banded exit-code contract.
+"""Smoke test for crawl_space.py's banded exit-code contract.
 
-Deterministic, network-free coverage: token-on-CLI rejection exits 1 without
-echoing the token; `--help` exits 0; banded constants + the `except Exception`
-(not BaseException) catch-all + the import guard are present (source). The
-runtime Tier2HardFail/HTTP mappings need a live keychain/server, so the
-catch-all's shape is asserted by source instead. Behavioral checks self-skip
-when deps aren't installed (the import guard exits 2 before main runs).
+Deterministic, network-free coverage: `--help` exits 0; banded constants + the
+`except Exception` (not BaseException) catch-all + the import guard + the
+preserved `130` SIGINT return are present (source). Runtime Tier2HardFail/HTTP
+mappings need a live keychain/server, so the catch-all's shape is asserted by
+source. The behavioral check self-skips when deps aren't installed.
 
 Run: python3 test_exit_codes.py     (exit 0 = all assertions pass)
 """
@@ -17,8 +16,11 @@ import pathlib
 import subprocess
 import sys
 
-HERE = pathlib.Path(__file__).resolve().parent
-CLI = HERE / "jira_align.py"
+HERE = pathlib.Path(__file__).resolve().parents[3] / ".apm" / "skills" \
+    / "confluence-crawler" / "scripts"
+if not HERE.is_dir():                     # wrong parents[] depth after a move
+    raise SystemExit(f"skill scripts dir not found at {HERE}")
+CLI = HERE / "crawl_space.py"
 SRC = CLI.read_text(encoding="utf-8")
 
 
@@ -44,13 +46,13 @@ def _run(*args: str) -> subprocess.CompletedProcess:
 
 def test_stdio_utf8_hardening_present() -> None:
     # Windows console hardening: stdout/stderr are reconfigured to UTF-8 at
-    # the top of file-path invocation, so non-ASCII output (UTF-8 JSON / CSV
-    # payloads on stdout — which is errors="strict" by default and *does*
-    # crash on a cp1252 console — plus em-dash messages) is safe. Must sit
-    # before the import guard so the guard's own messages are covered too.
-    # Source-asserted: a real non-UTF-8 Windows console isn't reproducible in
-    # a portable test (proven by A/B in the PR: --help crashed on `→`
-    # without this).
+    # the top of file-path invocation, so non-ASCII output (UTF-8 Markdown /
+    # JSON payloads on stdout — which is errors="strict" by default and
+    # *does* crash on a cp1252 console — plus em-dash messages) is safe. Must
+    # sit before the import guard so the guard's own messages are covered
+    # too. Source-asserted: a real non-UTF-8 Windows console isn't
+    # reproducible in a portable test (proven by A/B in the PR: --help
+    # crashed on `→` without this).
     assert 'reconfigure(encoding="utf-8")' in SRC, \
         "stdout/stderr UTF-8 hardening missing"
     assert SRC.index('if __package__ in (None, "") and __spec__ is None:') \
@@ -61,12 +63,14 @@ def test_stdio_utf8_hardening_present() -> None:
 
 
 def _deps_installed() -> bool:
-    proc = _run("--help")
-    return not (proc.returncode == 2 and "missing dependency" in proc.stderr)
+    # `--help` exits 0 only when every import resolves; any non-zero means the
+    # env can't run the CLI (a missing third-party dep), so skip behavioral.
+    return _run("--help").returncode == 0
 
 
 # CONVENTIONS § "The argv ban" canonical six — every credentialed CLI must
-# refuse these before argparse can echo the value.
+# refuse these before argparse can echo the value. The guard runs before
+# parse_args, so --check below is an incidental carrier (never consulted).
 CANONICAL_BANNED_FLAGS = (
     "--token", "--api-token", "--api-key", "--bearer", "--pat", "--password",
 )
@@ -77,7 +81,7 @@ def test_token_on_cli_rejected_exits_1_without_leak() -> None:
     # Exercised per-flag so a deny-set regression on any canonical flag
     # fails here, not just on --token.
     for flag in CANONICAL_BANNED_FLAGS:
-        proc = _run("check", flag, secret)
+        proc = _run("--check", flag, secret)
         assert proc.returncode == 1, f"{flag}: expected 1, got {proc.returncode}: {proc.stderr}"
         assert secret not in proc.stdout and secret not in proc.stderr, f"{flag}: token leaked"
         assert "must not be passed on the command line" in proc.stderr, \
@@ -86,9 +90,10 @@ def test_token_on_cli_rejected_exits_1_without_leak() -> None:
 
 def test_token_reject_wired_source() -> None:
     # Unconditional source check: the behavioral test above is in _BEHAVIORAL
-    # and self-skips when deps aren't installed (the import guard exits before
-    # main()'s reject runs) — i.e. the deps-less CI lint env. Guards the reject
-    # wiring and the canonical-six deny-set against silent drift.
+    # and self-skips when deps aren't installed (the import guard exits 2
+    # before main()'s reject runs) — which is exactly the deps-less CI lint
+    # env. Guards the reject wiring and the canonical-six deny-set against
+    # silent drift, where CI would otherwise see nothing.
     assert SRC.count("_reject_token_on_cli") >= 2, "reject helper not defined+called"
     for flag in CANONICAL_BANNED_FLAGS:
         assert f'"{flag}"' in SRC, f"canonical banned flag {flag} missing from deny-set"
@@ -100,7 +105,7 @@ def test_outofset_token_flag_value_scrubbed() -> None:
     # dotted JWT-shaped value also guards the regex charset (must include
     # `.`), per the security review.
     secret = "eyJhbGciOi." + "A" * 30 + ".Sig1234567890ABCDEF"  # noqa: S105 — JWT-shaped test literal
-    proc = _run("check", "--bogus-flag", secret)
+    proc = _run("--bogus-flag", secret)
     assert secret not in proc.stdout and secret not in proc.stderr, \
         "out-of-set token value leaked"
     assert "<scrubbed>" in proc.stderr, "value not scrubbed by the parser"
@@ -116,9 +121,11 @@ def test_banded_constants_defined() -> None:
         assert const in SRC, f"missing banded constant: {const}"
 
 
-def test_no_stale_constants() -> None:
-    for stale in ("EXIT_USER_ERROR", "EXIT_AUTH_ERROR", "EXIT_SERVER_ERROR"):
-        assert stale not in SRC, f"stale constant still present: {stale}"
+def test_no_bare_literal_returns_in_entry_path() -> None:
+    # The reconciled returns must use named constants, except the 130 SIGINT.
+    for bare in ("return 2\n", "return 0 if failed"):
+        assert bare not in SRC, f"bare-literal return still present: {bare!r}"
+    assert "return 130" in SRC, "the 130 SIGINT return must be preserved"
 
 
 def test_catch_all_is_except_exception_not_baseexception() -> None:
@@ -135,9 +142,9 @@ def test_import_guard_present() -> None:
 
 
 _BEHAVIORAL = {
+    "test_help_exits_0",
     "test_token_on_cli_rejected_exits_1_without_leak",
     "test_outofset_token_flag_value_scrubbed",
-    "test_help_exits_0",
 }
