+++ packs/atlassian/tests/skills/jira/test_auth_selector.py
