+++ packs/atlassian/tests/skills/jira/test_exit_codes.py
@@ -1,11 +1,17 @@
 #!/usr/bin/env python3
-"""Smoke test for crawl_space.py's banded exit-code contract.
+"""Smoke test for jira.py's banded exit-code contract.
 
-Deterministic, network-free coverage: `--help` exits 0; banded constants + the
-`except Exception` (not BaseException) catch-all + the import guard + the
-preserved `130` SIGINT return are present (source). Runtime Tier2HardFail/HTTP
-mappings need a live keychain/server, so the catch-all's shape is asserted by
-source. The behavioral check self-skips when deps aren't installed.
+Deterministic, network-free coverage:
+  - token-on-CLI rejection exits 1 (EXIT_ERROR) and never echoes the token;
+  - `--help` exits 0;
+  - the banded constants + the `except Exception` (not BaseException)
+    catch-all + the import guard are present (source inspection).
+
+The runtime Tier2HardFail/401/403/5xx mappings need a live keychain or server,
+which a standalone smoke test can't force deterministically; the catch-all's
+*shape* (except Exception → EXIT_ERROR, no str(exc) on the unexpected path) is
+asserted by source instead. If the skill's deps aren't installed, the behavioral
+checks self-skip (the import guard exits 2 before main runs).
 
 Run: python3 test_exit_codes.py     (exit 0 = all assertions pass)
 """
@@ -16,8 +22,11 @@ import pathlib
 import subprocess
 import sys
 
-HERE = pathlib.Path(__file__).resolve().parent
-CLI = HERE / "crawl_space.py"
+HERE = pathlib.Path(__file__).resolve().parents[3] / ".apm" / "skills" \
+    / "jira" / "scripts"
+if not HERE.is_dir():                     # wrong parents[] depth after a move
+    raise SystemExit(f"skill scripts dir not found at {HERE}")
+CLI = HERE / "jira.py"
 SRC = CLI.read_text(encoding="utf-8")
 
 
@@ -43,13 +52,13 @@ def _run(*args: str) -> subprocess.CompletedProcess:
 
 def test_stdio_utf8_hardening_present() -> None:
     # Windows console hardening: stdout/stderr are reconfigured to UTF-8 at
-    # the top of file-path invocation, so non-ASCII output (UTF-8 Markdown /
-    # JSON payloads on stdout — which is errors="strict" by default and
-    # *does* crash on a cp1252 console — plus em-dash messages) is safe. Must
-    # sit before the import guard so the guard's own messages are covered
-    # too. Source-asserted: a real non-UTF-8 Windows console isn't
-    # reproducible in a portable test (proven by A/B in the PR: --help
-    # crashed on `→` without this).
+    # the top of file-path invocation, so non-ASCII output (UTF-8 JSON / CSV
+    # payloads on stdout — which is errors="strict" by default and *does*
+    # crash on a cp1252 console — plus em-dash messages) is safe. Must sit
+    # before the import guard so the guard's own messages are covered too.
+    # Source-asserted: a real non-UTF-8 Windows console isn't reproducible in
+    # a portable test (proven by A/B in the PR: --help crashed on `→`
+    # without this).
     assert 'reconfigure(encoding="utf-8")' in SRC, \
         "stdout/stderr UTF-8 hardening missing"
     assert SRC.index('if __package__ in (None, "") and __spec__ is None:') \
@@ -60,14 +69,15 @@ def test_stdio_utf8_hardening_present() -> None:
 
 
 def _deps_installed() -> bool:
-    # `--help` exits 0 only when every import resolves; any non-zero means the
-    # env can't run the CLI (a missing third-party dep), so skip behavioral.
-    return _run("--help").returncode == 0
+    """False when the import guard fires (deps/shim absent) → skip behavioral."""
+    proc = _run("--help")
+    return not (proc.returncode == 2 and "missing dependency" in proc.stderr)
 
+
+# --- behavioral (real file-path invocation; need deps installed) -----------
 
 # CONVENTIONS § "The argv ban" canonical six — every credentialed CLI must
-# refuse these before argparse can echo the value. The guard runs before
-# parse_args, so --check below is an incidental carrier (never consulted).
+# refuse these before argparse can echo the value.
 CANONICAL_BANNED_FLAGS = (
     "--token", "--api-token", "--api-key", "--bearer", "--pat", "--password",
 )
@@ -78,7 +88,7 @@ def test_token_on_cli_rejected_exits_1_without_leak() -> None:
     # Exercised per-flag so a deny-set regression on any canonical flag
     # fails here, not just on --token.
     for flag in CANONICAL_BANNED_FLAGS:
-        proc = _run("--check", flag, secret)
+        proc = _run("check", flag, secret)
         assert proc.returncode == 1, f"{flag}: expected 1, got {proc.returncode}: {proc.stderr}"
         assert secret not in proc.stdout and secret not in proc.stderr, f"{flag}: token leaked"
         assert "must not be passed on the command line" in proc.stderr, \
@@ -87,10 +97,9 @@ def test_token_on_cli_rejected_exits_1_without_leak() -> None:
 
 def test_token_reject_wired_source() -> None:
     # Unconditional source check: the behavioral test above is in _BEHAVIORAL
-    # and self-skips when deps aren't installed (the import guard exits 2
-    # before main()'s reject runs) — which is exactly the deps-less CI lint
-    # env. Guards the reject wiring and the canonical-six deny-set against
-    # silent drift, where CI would otherwise see nothing.
+    # and self-skips when deps aren't installed (the import guard exits before
+    # main()'s reject runs) — i.e. the deps-less CI lint env. Guards the reject
+    # wiring and the canonical-six deny-set against silent drift.
     assert SRC.count("_reject_token_on_cli") >= 2, "reject helper not defined+called"
     for flag in CANONICAL_BANNED_FLAGS:
         assert f'"{flag}"' in SRC, f"canonical banned flag {flag} missing from deny-set"
@@ -102,7 +111,7 @@ def test_outofset_token_flag_value_scrubbed() -> None:
     # dotted JWT-shaped value also guards the regex charset (must include
     # `.`), per the security review.
     secret = "eyJhbGciOi." + "A" * 30 + ".Sig1234567890ABCDEF"  # noqa: S105 — JWT-shaped test literal
-    proc = _run("--bogus-flag", secret)
+    proc = _run("check", "--bogus-flag", secret)
     assert secret not in proc.stdout and secret not in proc.stderr, \
         "out-of-set token value leaked"
     assert "<scrubbed>" in proc.stderr, "value not scrubbed by the parser"
@@ -113,16 +122,16 @@ def test_help_exits_0() -> None:
     assert proc.returncode == 0, f"--help should exit 0, got {proc.returncode}"
 
 
+# --- structural (source inspection; always run) ----------------------------
+
 def test_banded_constants_defined() -> None:
     for const in ("EXIT_OK = 0", "EXIT_ERROR = 1", "EXIT_USER_ACTION = 2"):
         assert const in SRC, f"missing banded constant: {const}"
 
 
-def test_no_bare_literal_returns_in_entry_path() -> None:
-    # The reconciled returns must use named constants, except the 130 SIGINT.
-    for bare in ("return 2\n", "return 0 if failed"):
-        assert bare not in SRC, f"bare-literal return still present: {bare!r}"
-    assert "return 130" in SRC, "the 130 SIGINT return must be preserved"
+def test_no_stale_constants() -> None:
+    for stale in ("EXIT_USER_ERROR", "EXIT_AUTH_ERROR", "EXIT_SERVER_ERROR"):
+        assert stale not in SRC, f"stale constant still present: {stale}"
 
 
 def test_catch_all_is_except_exception_not_baseexception() -> None:
@@ -139,9 +148,9 @@ def test_import_guard_present() -> None:
 
 
 _BEHAVIORAL = {
-    "test_help_exits_0",
     "test_token_on_cli_rejected_exits_1_without_leak",
     "test_outofset_token_flag_value_scrubbed",
+    "test_help_exits_0",
 }
